QoL (#45)
I moved and added texts and buttons where I was supposed to add them. While testing and confirming the works, I noticed hearts and coins not resetting when going to main menu and starting the level anew. Made a function in game_manager to reset that that (for now only) activates when pressing the 'start game' button